refactor(random uuid): move get_namespace_and_name function to uuid.rs
